Extract logger
